jiexi2-master加很多工具/requirements.txt
absl-py==0.7.1
altgraph==0.16.1
arrow==0.14.5
astor==0.8.0
attrs==19.1.0
backports.lzma==0.0.13
beautifulsoup4==4.8.0
boto==2.49.0
boto3==1.9.199
botocore==1.12.199
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4

cycler==0.10.0
decorator==4.4.0

django-ftpserver==0.6.0

docutils==0.14
docx==0.2.4
docx2txt==0.8
feedparser==5.2.1
gast==0.2.2
gensim==3.8.0
grpcio==1.23.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
jieba==0.39
jmespath==0.9.4
JPype1==0.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
langid==1.1.6
lxml==4.4.0
Markdown==3.1.1
matplotlib==3.1.1
mock==3.0.5
networkx==2.3
ngender==0.1.1
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.25.0
pdfminer.six==20181108
pdfplumber==0.5.12
phone==0.4.2
phonenumbers==8.10.17
Pillow==6.1.0
protobuf==3.9.1
pudb==2019.1
pycryptodome==3.8.2
PyDocX==0.9.10
pyftpdlib==1.5.5
Pygments==2.4.2
pyhanlp==0.1.48
PyInstaller==3.5
pymongo==3.8.0
PyMySQL==0.9.3
pyparsing==2.4.2
pytesseract==0.2.7
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.2
PyWavelets==1.0.3
rarfile==3.0
reader==0.14
regex==2019.6.8
requests==2.22.0
s3transfer==0.2.1
scikit-image==0.15.0
scipy==1.3.0
seq2seq==0.0.5
six==1.12.0
smart-open==1.8.4
sortedcontainers==2.1.0
soupsieve==1.9.2
sqlparse==0.3.0
tensorboard==1.13.1

unicodecsv==0.14.1
urllib3==1.25.3
urwid==2.0.1
Wand==0.5.6
Werkzeug==0.15.5
wordninja==2.0.0
